✨ Each grabber does expose its own config class.
This begins a refactoring of the way grabbers are configured. The current way is just a Map<String,String> which is quite limited. I added a "namedParameters" but this way doesn't scale now does it?

So grabbers will give a class that will be filled in with reflection with whatever is in the config file.